+++ app/src/main/java/com/pict/pbl/medswift/screens/RegisterScreen.kt
@@ -1,0 +1,32 @@
+package com.pict.pbl.medswift.screens
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+
+class RegisterScreen : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            Surface(
+                modifier = Modifier.fillMaxSize(),
+                color = Color.White
+            ) {
+                ActivityUI()
+            }
+        }
+    }
+
+    @Composable
+    private fun ActivityUI() {
+        // TODO: Add register screen UI here
+    }
+
+
+}